+++ Makefile
@@ -33,6 +33,7 @@ ghcli_SRCS			=	src/ghcli.c			\
 					src/config.c			\
 					src/github/repos.c		\
 					src/github/comments.c		\
+					src/github/forks.c		\
 					thirdparty/sn/sn.c		\
 					thirdparty/pdjson/pdjson.c
