+++ src/forks.c
@@ -30,103 +30,25 @@
 #include <ghcli/config.h>
 #include <ghcli/curl.h>
 #include <ghcli/forks.h>
+#include <ghcli/github/forks.h>
 #include <ghcli/json_util.h>
 
 #include <pdjson/pdjson.h>
 
-static void
-parse_fork(struct json_stream *input, ghcli_fork *out)
-{
-    enum json_type  key_type   = JSON_NULL;
-    enum json_type  value_type = JSON_NULL;
-    const char     *key        = NULL;
-
-    if (json_next(input) != JSON_OBJECT)
-        errx(1, "Expected an object for a fork");
-
-    while ((key_type = json_next(input)) == JSON_STRING) {
-        size_t len;
-        key = json_get_string(input, &len);
-
-        if (strncmp("full_name", key, len) == 0) {
-            out->full_name = get_sv(input);
-        } else if (strncmp("owner", key, len) == 0) {
-            char *user = get_user(input);
-            out->owner = SV(user);
-        } else if (strncmp("created_at", key, len) == 0) {
-            out->date = get_sv(input);
-        } else if (strncmp("forks_count", key, len) == 0) {
-            out->forks = get_int(input);
-        } else {
-            value_type = json_next(input);
-
-            switch (value_type) {
-            case JSON_ARRAY:
-                json_skip_until(input, JSON_ARRAY_END);
-                break;
-            case JSON_OBJECT:
-                json_skip_until(input, JSON_OBJECT_END);
-                break;
-            default:
-                break;
-            }
-        }
-    }
-
-    if (key_type != JSON_OBJECT_END)
-        errx(1, "Fork object not closed");
-}
-
 int
 ghcli_get_forks(
     const char  *owner,
-    const char  *reponame,
+    const char  *repo,
     int          max,
     ghcli_fork **out)
 {
-    ghcli_fetch_buffer  buffer   = {0};
-    char               *url      = NULL;
-    char               *next_url = NULL;
-    enum   json_type    next     = JSON_NULL;
-    struct json_stream  stream   = {0};
-    int                 size     = 0;
-
-    *out = NULL;
-
-    url = sn_asprintf(
-        "%s/repos/%s/%s/forks",
-        ghcli_config_get_apibase(),
-        owner, reponame);
-
-    do {
-        ghcli_fetch(url, &next_url, &buffer);
-
-        json_open_buffer(&stream, buffer.data, buffer.length);
-        json_set_streaming(&stream, 1);
-
-        // TODO: Poor error message
-        if ((next = json_next(&stream)) != JSON_ARRAY)
-            errx(1,
-                 "Expected array in response from API "
-                 "but got something else instead");
-
-        while ((next = json_peek(&stream)) != JSON_ARRAY_END) {
-            *out = realloc(*out, sizeof(ghcli_fork) * (size + 1));
-            ghcli_fork *it = &(*out)[size++];
-            parse_fork(&stream, it);
-
-            if (size == max)
-                break;
-        }
-
-        json_close(&stream);
-        free(buffer.data);
-        free(url);
-    } while ((url = next_url) && (max == -1 || size < max));
-
-    free(next_url);
-
-    return size;
+    switch (ghcli_config_get_forge_type()) {
+    case GHCLI_FORGE_GITHUB:
+        return github_get_forks(owner, repo, max, out);
+    default:
+        sn_unimplemented;
+    }
+    return -1;
 }
 
 void
@@ -170,26 +92,12 @@ ghcli_print_forks(
 void
 ghcli_fork_create(const char *owner, const char *repo, const char *_in)
 {
-    char               *url       = NULL;
-    char               *post_data = NULL;
-    sn_sv               in        = SV_NULL;
-    ghcli_fetch_buffer  buffer    = {0};
-
-    url = sn_asprintf(
-        "%s/repos/%s/%s/forks",
-        ghcli_config_get_apibase(),
-        owner, repo);
-    if (_in) {
-        in        = ghcli_json_escape(SV((char *)_in));
-        post_data = sn_asprintf("{\"organization\":\""SV_FMT"\"}",
-                                SV_ARGS(in));
+    switch (ghcli_config_get_forge_type()) {
+    case GHCLI_FORGE_GITHUB:
+        github_fork_create(owner, repo, _in);
+        break;
+    default:
+        sn_unimplemented;
+        break;
     }
-
-    ghcli_fetch_with_method("POST", url, post_data, NULL, &buffer);
-    ghcli_print_html_url(buffer);
-
-    free(in.data);
-    free(url);
-    free(post_data);
-    free(buffer.data);
 }
